Use non-blocking CUDA streams in tests.
Blocking streams are being synchronized each time an action is
performed on the default
stream (cf. https://docs.nvidia.com/cuda/cuda-runtime-api/stream-sync-behavior.html#stream-sync-behavior).
Therefore, let us not assume this synchronization in our tests, and
use non-blocking streams instead. This requires manually synchronizing
a stream between a call to `recv()` and a `cudaMemcpy()` (which, in
the non-async version happens on the default stream).

ghstack-source-id: 3357b5307af48f9b7dcf3ef1b2575ab3682fefa2
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/279